+++ CHANGES.md
@@ -3,6 +3,7 @@
 
 ## v5.0.0. (2023.Q3)
  - Switch to binary transfer mode 
+ - Added WebSocket Heartbeat
  - New optimized Quark data stream
  - AES key derivation use
  - Performance optimizations